ocr-service/requirements.txt
# Framework
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.11.*
pydantic-settings==2.8.*

# --- VLM: Qwen3-VL-8B (primary backend) ---
torch==2.6.*
transformers==4.50.*
accelerate==1.5.*
bitsandbytes==0.45.*
qwen-vl-utils==0.0.*
Pillow==11.*

# --- PaddlePaddle (legacy fallback) ---
paddlepaddle==3.0.*
paddleocr==2.9.*
paddlex==3.0.*

# PDF Processing
PyMuPDF==1.25.*
pdf2image==1.17.*

# Image Processing
opencv-python-headless==4.10.*

# Utilities
numpy==2.2.*
httpx==0.28.*
python-multipart==0.0.*
orjson==3.10.*
boto3==1.37.*
